Renamed setup files
